pkg/tarheader: cast inode up to uint64
Some systems (FreeBSD included) have uint32 as syscall.Stat_t.Ino

+++ pkg/tarheader/pop_posix.go
@@ -18,11 +18,11 @@ func populateHeaderUnix(h *tar.Header, fi os.FileInfo, seen map[uint64]string) {
 	h.Uid = int(st.Uid)
 	h.Gid = int(st.Gid)
 	// If we have already seen this inode, generate a hardlink
-	p, ok := seen[st.Ino]
+	p, ok := seen[uint64(st.Ino)]
 	if ok {
 		h.Linkname = p
 		h.Typeflag = tar.TypeLink
 	} else {
-		seen[st.Ino] = h.Name
+		seen[uint64(st.Ino)] = h.Name
 	}
 }
